[2.3.x] Large cleanup of minor code issues.

+++ modules/grizzly/src/main/java/org/glassfish/grizzly/compression/lzma/impl/lz/BinTree.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2011 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011-2015 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -164,11 +164,11 @@ public class BinTree extends InWindow {
         if (HASH_ARRAY) {
             int temp = CrcTable[_bufferBase[cur] & 0xFF] ^ (_bufferBase[cur + 1] & 0xFF);
             hash2Value = temp & (kHash2Size - 1);
-            temp ^= ((int) (_bufferBase[cur + 2] & 0xFF) << 8);
+            temp ^= ((_bufferBase[cur + 2] & 0xFF) << 8);
             hash3Value = temp & (kHash3Size - 1);
             hashValue = (temp ^ (CrcTable[_bufferBase[cur + 3] & 0xFF] << 5)) & _hashMask;
         } else {
-            hashValue = ((_bufferBase[cur] & 0xFF) ^ ((int) (_bufferBase[cur + 1] & 0xFF) << 8));
+            hashValue = ((_bufferBase[cur] & 0xFF) ^ ((_bufferBase[cur + 1] & 0xFF) << 8));
         }
 
         int curMatch = _hash[kFixHashSize + hashValue];
@@ -283,12 +283,12 @@ public class BinTree extends InWindow {
                 int temp = CrcTable[_bufferBase[cur] & 0xFF] ^ (_bufferBase[cur + 1] & 0xFF);
                 int hash2Value = temp & (kHash2Size - 1);
                 _hash[hash2Value] = _pos;
-                temp ^= ((int) (_bufferBase[cur + 2] & 0xFF) << 8);
+                temp ^= ((_bufferBase[cur + 2] & 0xFF) << 8);
                 int hash3Value = temp & (kHash3Size - 1);
                 _hash[kHash3Offset + hash3Value] = _pos;
                 hashValue = (temp ^ (CrcTable[_bufferBase[cur + 3] & 0xFF] << 5)) & _hashMask;
             } else {
-                hashValue = ((_bufferBase[cur] & 0xFF) ^ ((int) (_bufferBase[cur + 1] & 0xFF) << 8));
+                hashValue = ((_bufferBase[cur] & 0xFF) ^ ((_bufferBase[cur + 1] & 0xFF) << 8));
             }
 
             int curMatch = _hash[kFixHashSize + hashValue];
