+++ modules/grizzly/src/main/java/org/glassfish/grizzly/utils/StringDecoder.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2009-2011 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2009-2015 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -137,7 +137,7 @@ public class StringDecoder extends AbstractTransformer<Buffer, String> {
                 return TransformationResult.createIncompletedResult(input);
             }
 
-            stringSize = (int) input.getInt();
+            stringSize = input.getInt();
             lengthAttribute.set(storage, stringSize);
         }
