+++ modules/http-ajp/src/main/java/org/glassfish/grizzly/http/ajp/AjpHandlerFilter.java
@@ -499,7 +499,7 @@ public class AjpHandlerFilter extends BaseFilter {
             buffer.put((byte) 'B');
             buffer.putShort((short) 3);
             buffer.put(AjpConstants.JK_AJP13_GET_BODY_CHUNK);
-            buffer.putShort((short) AjpConstants.SUGGESTED_MAX_PAYLOAD_SIZE);
+            buffer.putShort(AjpConstants.SUGGESTED_MAX_PAYLOAD_SIZE);
 
             buffer.flip();
             buffer.allowBufferDispose(true);
