+++ modules/http-server/src/main/java/org/glassfish/grizzly/http/server/accesslog/AccessLogProbe.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2014 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014-2015 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -140,7 +140,7 @@ public class AccessLogProbe extends HttpServerProbe.Adapter {
         final long timeStamp = System.currentTimeMillis();
         final long nanoStamp = System.nanoTime();
 
-        final long responseNanos = requestNanos == null ? -1 : nanoStamp - requestNanos.longValue();
+        final long responseNanos = requestNanos == null ? -1 : nanoStamp - requestNanos;
         final Date requestMillis = new Date(timeStamp - (responseNanos / 1000000000L));
 
         /* Create a formatted log entry string and append it */
