+++ modules/http/src/main/java/org/glassfish/grizzly/http/util/Ascii.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2010-2014 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2010-2015 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -95,7 +95,7 @@ public final class Ascii {
      */
     public static void toUpper(final byte[] bytes) {
         for (int i = 0; i < bytes.length; i++) {
-            byte b = (byte)bytes[i];
+            byte b = bytes[i];
             bytes[i] = (byte)((b >= 'a' && b <= 'z') ? (b + 'A' - 'a') : b);
         }
     }
@@ -113,7 +113,7 @@ public final class Ascii {
      */
     public static void toLower(final byte[] bytes) {
         for (int i = 0; i < bytes.length; i++) {
-            byte b = (byte)bytes[i];
+            byte b = bytes[i];
             bytes[i] = (byte)((b >= 'A' && b <= 'Z') ? (b - 'A' + 'a') : b);
         }
     }
